can display sth user message after scanning qr

+++ restdemo-master/src/main/java/com/example/demo/model/Checkin.java
@@ -16,10 +16,10 @@ public class Checkin {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "checkin_id")
-    private Long checkinId;
+    private int checkinId;
 
     @Column(name = "registration_id", nullable = false, unique = true)
-    private Long registrationId;
+    private int registrationId;
 
     @Column(name = "checkin_time", nullable = false)
     private LocalDateTime checkinTime;
@@ -28,25 +28,25 @@ public class Checkin {
     public Checkin() {
     }
 
-    public Checkin(Long registrationId, LocalDateTime checkinTime) {
+    public Checkin(int registrationId, LocalDateTime checkinTime) {
         this.registrationId = registrationId;
         this.checkinTime = checkinTime;
     }
 
     // Getters and Setters
-    public Long getCheckinId() {
+    public int getCheckinId() {
         return checkinId;
     }
 
-    public void setCheckinId(Long checkinId) {
+    public void setCheckinId(int checkinId) {
         this.checkinId = checkinId;
     }
 
-    public Long getRegistrationId() {
+    public int getRegistrationId() {
         return registrationId;
     }
 
-    public void setRegistrationId(Long registrationId) {
+    public void setRegistrationId(int registrationId) {
         this.registrationId = registrationId;
     }
